refactor: remove `new()` from DbManager
Different database backend might require quite different data and struct
to create a manager. It causes more harm than good to hard code a struct
used to create a db manager.
